refactors util functions
